requirements.txt
Cheetah==2.4.4
Django==1.10.3
Landscape-Client==14.12
Markdown==2.6.7
PAM==0.4.2
PyYAML==3.10
SecretStorage==2.0.0
Twisted-Core==13.2.0
Twisted-Names==13.2.0
Twisted-Web==13.2.0
apt-xapian-index==0.45
argparse==1.2.1
chardet==2.0.1
cloud-init==0.7.5
colorama==0.2.5
configobj==4.7.2
django-multiselectfield==0.1.4
html5lib==0.999
httplib2==0.8
jsonpatch==1.3
jsonpointer==1.0
keyring==3.5
launchpadlib==1.10.2
lazr.restfulclient==0.13.3
lazr.uri==1.0.3
oauth==1.0.1
pbr==1.10.0
prettytable==0.7.2
psycopg2==2.6.2
pyOpenSSL==0.13
pycrypto==2.6.1
pycurl==7.19.3
pygobject==3.12.0
pyserial==2.6
python-apt==0.9.3.5ubuntu2
python-debian==0.1.21-nmu2ubuntu2
requests==2.2.1
simplejson==3.3.1
six==1.5.2
ssh-import-id==3.21
testresources==2.0.1
urllib3==1.7.1
wadllib==1.3.2
wheel==0.24.0
wsgiref==0.1.2
zope.interface==4.0.5
